// src/app/modules/academics/student-information-system/model/group-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StudentGroupDetailsView
 */
export interface GroupView {
    studentGroupId: string;
    representatives?: Array<string>;
    staffName?: Array<string>;
    studentName?: Array<string>;
    schoolAcademicSessionId?: any;
    sortBy: string;
    sortOrder?: number;
    pageNumber: number;
    pageSize: number;
}